Extract channel form to separate hook
